requirements.txt
fastapi[all]
uvicorn[standard]
gunicorn
requests
structlog
circuitbreaker
prometheus_client
tenacity
PyYAML
tenacity